Use more inner classes
