+++ matsim/src/main/java/org/matsim/core/population/io/ParallelPopulationWriterHandlerV6.java
@@ -1,14 +1,22 @@
 package org.matsim.core.population.io;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.matsim.api.core.v01.population.Person;
-import org.matsim.api.core.v01.population.Population;
+import org.matsim.api.core.v01.Coord;
+import org.matsim.api.core.v01.Id;
+import org.matsim.api.core.v01.network.Link;
+import org.matsim.api.core.v01.population.*;
+import org.matsim.core.population.PersonUtils;
+import org.matsim.core.population.routes.NetworkRoute;
+import org.matsim.core.router.TripStructureUtils;
 import org.matsim.core.utils.geometry.CoordinateTransformation;
 import org.matsim.core.utils.io.MatsimXmlWriter;
 import org.matsim.core.utils.misc.Counter;
+import org.matsim.core.utils.misc.Time;
 import org.matsim.utils.objectattributes.AttributeConverter;
+import org.matsim.utils.objectattributes.attributable.Attributes;
+import org.matsim.utils.objectattributes.attributable.AttributesImpl;
+import org.matsim.utils.objectattributes.attributable.AttributesUtils;
 import org.matsim.utils.objectattributes.attributable.AttributesXmlWriterDelegate;
+import org.matsim.vehicles.Vehicle;
 
 import java.io.BufferedWriter;
 import java.io.IOException;
@@ -20,12 +28,12 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.LinkedBlockingQueue;
 
 import static org.matsim.core.utils.io.XmlUtils.encodeAttributeValue;
+import static org.matsim.core.utils.io.XmlUtils.encodeContent;
 
 /**
  * @author steffenaxer
  */
 public class ParallelPopulationWriterHandlerV6 implements PopulationWriterHandler{
-	static final Logger LOG = LogManager.getLogger(ParallelPopulationWriterHandlerV6.class);
 	private static final int THREAD_LIMIT = 2;
 	private static final int MAX_QUEUE_LENGTH = 1000;
 	private final BlockingQueue<PersonData> inputQueue = new LinkedBlockingQueue<>();
@@ -34,7 +42,7 @@ public class ParallelPopulationWriterHandlerV6 implements PopulationWriterHandle
 	private final AttributesXmlWriterDelegate attributesWriter = new AttributesXmlWriterDelegate();
 	private Thread[] threads;
 	private Thread writeThread;
-	private ParallelPopulationCreatorV6[] runners;
+	private PersonStringCreator[] runners;
 	private ParallelPopulationWriterV6 writer;
 	private final Map<Class<?>, AttributeConverter<?>> converters = new HashMap<>();
 
@@ -48,11 +56,11 @@ public class ParallelPopulationWriterHandlerV6 implements PopulationWriterHandle
 		{
 			int computeThreads = Math.min(THREAD_LIMIT,Runtime.getRuntime().availableProcessors());
 			threads = new Thread[computeThreads];
-			runners = new ParallelPopulationCreatorV6[computeThreads];
+			runners = new PersonStringCreator[computeThreads];
 			for (int i = 0; i < computeThreads; i++) {
 
-				ParallelPopulationCreatorV6 runner =
-						new ParallelPopulationCreatorV6(
+				PersonStringCreator runner =
+						new PersonStringCreator(
 								this.coordinateTransformation,
 								this.inputQueue);
 
@@ -60,7 +68,7 @@ public class ParallelPopulationWriterHandlerV6 implements PopulationWriterHandle
 				this.runners[i] = runner;
 				Thread thread = new Thread(runner);
 				thread.setDaemon(true);
-				thread.setName(ParallelPopulationCreatorV6.class.toString() + i);
+				thread.setName(PersonStringCreator.class.toString() + i);
 				threads[i] = thread;
 				thread.start();
 			}
@@ -71,7 +79,7 @@ public class ParallelPopulationWriterHandlerV6 implements PopulationWriterHandle
 	{
 		if(this.writeThread == null)
 		{
-			this.writer = new ParallelPopulationWriterV6(this.outputQueue,out);
+			this.writer = new ParallelPopulationWriterV6(this.outputQueue, out);
 			writeThread = new Thread(this.writer);
 			writeThread.setDaemon(true);
 			writeThread.setName(ParallelPopulationWriterV6.class.toString() + 0);
@@ -80,7 +88,7 @@ public class ParallelPopulationWriterHandlerV6 implements PopulationWriterHandle
 
 	}
 
-	private void initObjectAttributeConverters(ParallelPopulationCreatorV6 runner)
+	private void initObjectAttributeConverters(PersonStringCreator runner)
 	{
 		runner.putAttributeConverters(this.converters);
 	}
@@ -164,10 +172,10 @@ public class ParallelPopulationWriterHandlerV6 implements PopulationWriterHandle
 
 	public record PersonData(Person person, CompletableFuture<String> futurePersonString){}
 
-	public class ParallelPopulationWriterV6 implements Runnable {
+	public static class ParallelPopulationWriterV6 implements Runnable {
 		private final Counter counter = new Counter("[" + this.getClass().getSimpleName() + "] dumped person # ");
-		private BlockingQueue<CompletableFuture<String>> outputQueue;
-		private BufferedWriter out;
+		private final BlockingQueue<CompletableFuture<String>> outputQueue;
+		private final BufferedWriter out;
 
 		private boolean finish = false;
 
@@ -195,6 +203,255 @@ public class ParallelPopulationWriterHandlerV6 implements PopulationWriterHandle
 
 		public void finish() {
 			this.finish = true;
+		}
+	}
+
+	static class PersonStringCreator implements Runnable {
+		private final AttributesXmlWriterDelegate attributesWriter = new AttributesXmlWriterDelegate();
+		private final CoordinateTransformation coordinateTransformation;
+		private final BlockingQueue<ParallelPopulationWriterHandlerV6.PersonData> queue;
+		private final StringBuilder routeDescStringBuilder = new StringBuilder(100);
+		private final StringBuilder stringBuilder = new StringBuilder(100_000);
+		private boolean finish = false;
+
+		PersonStringCreator(CoordinateTransformation coordinateTransformation, BlockingQueue<ParallelPopulationWriterHandlerV6.PersonData> queue) {
+			this.coordinateTransformation = coordinateTransformation;
+			this.queue = queue;
+		}
+
+		private static void endPerson(final StringBuilder out)  {
+			out.append("\t</person>\n\n");
+		}
+
+		private static void endPlan(final StringBuilder out) {
+			out.append("\t\t</plan>\n\n");
+		}
+
+		private static void endLeg(final StringBuilder out)  {
+			out.append("\t\t\t</leg>\n");
+		}
+
+		private void startRoute(final Route route, final StringBuilder out) {
+			out.append("\t\t\t\t<route ");
+			out.append("type=\"");
+			out.append(encodeAttributeValue(route.getRouteType()));
+			out.append("\"");
+			out.append(" start_link=\"");
+			out.append(encodeAttributeValue(route.getStartLinkId().toString()));
+			out.append("\"");
+			out.append(" end_link=\"");
+			out.append(encodeAttributeValue(route.getEndLinkId().toString()));
+			out.append("\"");
+			out.append(" trav_time=\"");
+			out.append(Time.writeTime(route.getTravelTime()));
+			out.append("\"");
+			out.append(" distance=\"");
+			out.append(route.getDistance());
+			out.append("\"");
+			if (route instanceof NetworkRoute networkRoute) {
+				out.append(" vehicleRefId=\"");
+				final Id<Vehicle> vehicleId = networkRoute.getVehicleId();
+				if (vehicleId == null) {
+					out.append("null");
+				} else {
+					out.append(encodeAttributeValue(vehicleId.toString()));
+				}
+				out.append("\"");
+			}
+			out.append(">");
+
+			String rd;
+			if (route instanceof NetworkRoute networkRoute) {
+				rd = getRouteDescriptionNetworkRoute(this.routeDescStringBuilder, networkRoute);
+				this.routeDescStringBuilder.setLength(0);
+			} else {
+				rd = route.getRouteDescription();
+			}
+			if (rd != null) {
+				out.append(encodeContent(rd));
+			}
+		}
+
+		private static void endRoute(final StringBuilder out) {
+			out.append("</route>\n");
+		}
+
+		public void putAttributeConverters(final Map<Class<?>, AttributeConverter<?>> converters) {
+			this.attributesWriter.putAttributeConverters(converters);
+		}
+
+		private void process() throws IOException {
+			do {
+				ParallelPopulationWriterHandlerV6.PersonData personData = this.queue.poll();
+				if (personData != null) {
+					Person person = personData.person();
+					StringBuilder out = stringBuilder;
+
+					this.startPerson(person, out);
+					for (Plan plan : person.getPlans()) {
+						startPlan(plan, out);
+						// act/leg
+						for (PlanElement pe : plan.getPlanElements()) {
+							if (pe instanceof Activity act) {
+								this.writeAct(act, out);
+							} else if (pe instanceof Leg leg) {
+								this.startLeg(leg, out);
+								// route
+								Route route = leg.getRoute();
+								if (route != null) {
+									startRoute(route, out);
+									endRoute(out);
+								}
+								endLeg(out);
+							}
+						}
+						endPlan(out);
+					}
+					endPerson(out);
+					this.writeSeparator(out);
+					CompletableFuture<String> completableFuture = personData.futurePersonString();
+					completableFuture.complete(out.toString());
+
+					// Reset stringBuilder instead of instantiate
+					stringBuilder.setLength(0);
+				}
+			} while (!(this.queue.isEmpty() && finish));
+		}
+
+		public void finish() {
+			this.finish = true;
+		}
+
+		private void startPerson(final Person person, final StringBuilder out) {
+			out.append("\t<person id=\"");
+			out.append(encodeAttributeValue(person.getId().toString()));
+			out.append("\"");
+			out.append(">\n");
+			this.attributesWriter.writeAttributes("\t\t", out, person.getAttributes());
+		}
+
+		private void startPlan(final Plan plan, final StringBuilder out)  {
+			out.append("\t\t<plan");
+			if (plan.getScore() != null) {
+				out.append(" score=\"");
+				out.append(plan.getScore().toString());
+				out.append("\"");
+			}
+			if (PersonUtils.isSelected(plan))
+				out.append(" selected=\"yes\"");
+			else
+				out.append(" selected=\"no\"");
+			if ((plan.getType() != null)) {
+				out.append(" type=\"");
+				out.append(encodeAttributeValue(plan.getType()));
+				out.append("\"");
+			}
+			out.append(">\n");
+
+			this.attributesWriter.writeAttributes("\t\t\t\t", out, plan.getAttributes());
+		}
+
+		private void writeAct(final Activity act, final StringBuilder out) {
+			out.append("\t\t\t<activity type=\"");
+			out.append(encodeAttributeValue(act.getType()));
+			out.append("\"");
+			if (act.getLinkId() != null) {
+				out.append(" link=\"");
+				out.append(encodeAttributeValue(act.getLinkId().toString()));
+				out.append("\"");
+			}
+			if (act.getFacilityId() != null) {
+				out.append(" facility=\"");
+				out.append(encodeAttributeValue(act.getFacilityId().toString()));
+				out.append("\"");
+			}
+			if (act.getCoord() != null) {
+				final Coord coord = this.coordinateTransformation.transform(act.getCoord());
+				out.append(" x=\"");
+				out.append(coord.getX());
+				out.append("\" y=\"");
+				out.append(coord.getY());
+				out.append("\"");
+
+				if (act.getCoord().hasZ()) {
+					out.append(" z=\"");
+					out.append(coord.getZ());
+					out.append("\"");
+				}
+			}
+			if (act.getStartTime().isDefined()) {
+				out.append(" start_time=\"");
+				out.append(Time.writeTime(act.getStartTime().seconds()));
+				out.append("\"");
+			}
+			if (act.getMaximumDuration().isDefined()) {
+				out.append(" max_dur=\"");
+				out.append(Time.writeTime(act.getMaximumDuration().seconds()));
+				out.append("\"");
+			}
+			if (act.getEndTime().isDefined()) {
+				out.append(" end_time=\"");
+				out.append(Time.writeTime(act.getEndTime().seconds()));
+				out.append("\"");
+			}
+			out.append(" >\n");
+
+			this.attributesWriter.writeAttributes("\t\t\t\t", out, act.getAttributes());
+
+			out.append("\t\t\t</activity>\n");
+		}
+
+		private void startLeg(final Leg leg, final StringBuilder out) {
+			out.append("\t\t\t<leg mode=\"");
+			out.append(encodeAttributeValue(leg.getMode()));
+			out.append("\"");
+			if (leg.getDepartureTime().isDefined()) {
+				out.append(" dep_time=\"");
+				out.append(Time.writeTime(leg.getDepartureTime().seconds()));
+				out.append("\"");
+			}
+			if (leg.getTravelTime().isDefined()) {
+				out.append(" trav_time=\"");
+				out.append(Time.writeTime(leg.getTravelTime().seconds()));
+				out.append("\"");
+			}
+
+			out.append(">\n");
+
+			if (leg.getRoutingMode() != null) {
+				Attributes attributes = new AttributesImpl();
+				AttributesUtils.copyTo(leg.getAttributes(), attributes);
+				attributes.putAttribute(TripStructureUtils.routingMode, leg.getRoutingMode());
+				this.attributesWriter.writeAttributes("\t\t\t\t", out, attributes);
+			} else this.attributesWriter.writeAttributes("\t\t\t\t", out, leg.getAttributes());
+		}
+
+		public void writeSeparator(final StringBuilder out) throws IOException {
+			out.append("<!-- ====================================================================== -->\n\n");
+		}
+
+		@Override
+		public void run() {
+			try {
+				this.process();
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
+		}
+
+		// This method re-uses the existing StringBuilder
+		public static String getRouteDescriptionNetworkRoute(StringBuilder desc, NetworkRoute route) {
+			desc.append(route.getStartLinkId().toString());
+			for (Id<Link> linkId : route.getLinkIds()) {
+				desc.append(" ");
+				desc.append(linkId.toString());
+			}
+			// If the start links equals the end link additionally check if its is a round trip.
+			if (!route.getEndLinkId().equals(route.getStartLinkId()) || route.getLinkIds().size() > 0) {
+				desc.append(" ");
+				desc.append(route.getEndLinkId().toString());
+			}
+			return desc.toString();
 		}
 	}
 }
